+++ Bienestar.Negocio/DTO/UsuarioDTO.cs
@@ -4,7 +4,7 @@ namespace Bienestar.Negocio.DTO
     public class UsuarioDTO
     {
         public int Id { get; set; }
-        public int NumeroIdentificacion { get; set; }
+        public Int64 NumeroIdentificacion { get; set; }
         public string Nombres { get; set; }
         public string Apellidos { get; set; }
         public DateTime FechaNacimiento { get; set; }
